requirements.txt
DBUtils
Flask
pyMySQL
requests
Jinja2
Flask-SocketIO
Pillow